+++ frontend/src/components/editor-x/editor.tsx
@@ -1,0 +1,73 @@
+"use client";
+
+import { InitialConfigType, LexicalComposer } from "@lexical/react/LexicalComposer";
+import { OnChangePlugin } from "@lexical/react/LexicalOnChangePlugin";
+import { EditorState, SerializedEditorState } from "lexical";
+
+import { editorTheme } from "@/components/ui/editor/themes/editor-theme";
+import { TooltipProvider } from "@/components/ui/tooltip";
+
+import { nodes } from "./nodes";
+import { Plugins } from "./plugins";
+import { cn } from "@/lib/utils";
+import type { UserPublic } from "@/types/api";
+
+const editorConfig: InitialConfigType = {
+  namespace: "Editor",
+  theme: editorTheme,
+  nodes,
+  onError: (error: Error) => {
+    console.error(error);
+  },
+};
+
+export function Editor({
+  editorState,
+  editorSerializedState,
+  onChange,
+  onSerializedChange,
+  readOnly = false,
+  showToolbar = true,
+  className,
+  mentionableUsers = [],
+}: {
+  editorState?: EditorState;
+  editorSerializedState?: SerializedEditorState;
+  onChange?: (editorState: EditorState) => void;
+  onSerializedChange?: (editorSerializedState: SerializedEditorState) => void;
+  readOnly?: boolean;
+  showToolbar?: boolean;
+  className?: string;
+  mentionableUsers?: UserPublic[];
+}) {
+  return (
+    <div className={cn("bg-background overflow-y-auto rounded-lg border shadow", className)}>
+      <LexicalComposer
+        initialConfig={{
+          ...editorConfig,
+          editable: !readOnly,
+          ...(editorState ? { editorState } : {}),
+          ...(editorSerializedState ? { editorState: JSON.stringify(editorSerializedState) } : {}),
+        }}
+      >
+        <TooltipProvider>
+          <Plugins
+            showToolbar={showToolbar}
+            readOnly={readOnly}
+            mentionableUsers={mentionableUsers}
+          />
+
+          {!readOnly && (
+            <OnChangePlugin
+              ignoreSelectionChange={true}
+              onChange={(editorState) => {
+                onChange?.(editorState);
+                onSerializedChange?.(editorState.toJSON());
+              }}
+            />
+          )}
+        </TooltipProvider>
+      </LexicalComposer>
+    </div>
+  );
+}