+++ frontend/src/components/ui/editor/plugins/link-plugin.tsx
@@ -1,0 +1,8 @@
+import { JSX } from "react";
+import { LinkPlugin as LexicalLinkPlugin } from "@lexical/react/LexicalLinkPlugin";
+
+import { validateUrl } from "@/components/ui/editor/utils/url";
+
+export function LinkPlugin(): JSX.Element {
+  return <LexicalLinkPlugin validateUrl={validateUrl} />;
+}